+++ database/seed.js
@@ -1,5 +1,7 @@
 const mongoose = require('mongoose');
-// const profileImages = require('./dbHelpler.js/index.js');
+const {profileImages} = require('./dbHelper.js');
+console.log('profile images', typeof profileImages);
+const productList = require('./sample_products.js');
 mongoose.connect('mongodb://localhost/Images', {useNewUrlParser: true, useUnifiedTopology: true});
 
 var db = mongoose.connection;
@@ -11,25 +13,33 @@ db.on('open', () => {
   });
 });
 
+const productListWithImages = productList.map((product) => {
+  const productImages = profileImages(['/image1', '/image2', '/image3', '/image4', '/image5', '/image6', '/image7', '/image8', '/image9', '/image10', '/image1', 'image11', '/image12', '/image13', '/image14', '/image15'], 10);
+  product['imageUrls'] = productImages;
+  return product;
+});
+
 db.once('open', function() {
   console.log('we\'re connected!');
 
   var imageSchema = new mongoose.Schema({
     productId: String,
-    imageUrl: String
+    productName: String,
+    imageUrls: {
+      type: Array,
+      default: undefined
+    }
   });
 
   var Image = mongoose.model('Image', imageSchema);
 
-  var product1 = new Image({'productId': '0001222', 'imageUrl':'/new/imageUrl'});
-
-  product1.save(function (err, product) {
-    if (err) {
-      return console.error(err);
-    }
-    console.log('product', product);
-    db.close();
-
-  });
-
+  Image.insertMany(productListWithImages, (err, response) => {
+      if (err) {
+        console.log('error', error);
+      } else {
+        console.log('db insertion complete', response);
+        db.close();
+      }
+   }
+  );
 });
